fix: Fix 3 remaining failing SQL examples
- m5_changepoint_examples.sql: Cast UBIGINT cp_index to INTEGER for array indexing
- synthetic_forecasting_examples.sql: Add json extension, use ts_forecast_exog_by
- synthetic_period_examples.sql: Rewrite with ts_detect_periods_by table macro

All 16 example files now pass validation.

+++ examples/changepoint_detection/m5_changepoint_examples.sql
@@ -265,7 +265,7 @@ cp_unnested AS (
 SELECT
     c.item_id,
     c.cp_index,
-    d.date_list[c.cp_index + 1] AS cp_date  -- +1 because DuckDB arrays are 1-indexed
+    d.date_list[(c.cp_index + 1)::INTEGER] AS cp_date  -- +1 because DuckDB arrays are 1-indexed
 FROM cp_unnested c
 JOIN item_dates d ON c.item_id = d.item_id;
